flesh out local for demo
