+++ apps/plugins/pkg/app/meta/manager_test.go
@@ -35,10 +35,12 @@ func TestProviderManager_GetMeta(t *testing.T) {
 	ctx := context.Background()
 
 	t.Run("returns cached result when available and not expired", func(t *testing.T) {
-		cachedMeta := pluginsv0alpha1.PluginMetaJSONData{
-			Id:   "test-plugin",
-			Name: "Test Plugin",
-			Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+		cachedMeta := pluginsv0alpha1.PluginMetaSpec{
+			PluginJSON: pluginsv0alpha1.PluginMetaJSONData{
+				Id:   "test-plugin",
+				Name: "Test Plugin",
+				Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+			},
 		}
 
 		provider := &mockProvider{
@@ -60,8 +62,10 @@ func TestProviderManager_GetMeta(t *testing.T) {
 
 		provider.getMetaFunc = func(ctx context.Context, pluginID, version string) (*Result, error) {
 			return &Result{
-				Meta: pluginsv0alpha1.PluginMetaJSONData{Id: "different"},
-				TTL:  time.Hour,
+				Meta: pluginsv0alpha1.PluginMetaSpec{
+					PluginJSON: pluginsv0alpha1.PluginMetaJSONData{Id: "different"},
+				},
+				TTL: time.Hour,
 			}, nil
 		}
 
@@ -73,10 +77,12 @@ func TestProviderManager_GetMeta(t *testing.T) {
 	})
 
 	t.Run("fetches from provider when not cached", func(t *testing.T) {
-		expectedMeta := pluginsv0alpha1.PluginMetaJSONData{
-			Id:   "test-plugin",
-			Name: "Test Plugin",
-			Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+		expectedMeta := pluginsv0alpha1.PluginMetaSpec{
+			PluginJSON: pluginsv0alpha1.PluginMetaJSONData{
+				Id:   "test-plugin",
+				Name: "Test Plugin",
+				Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+			},
 		}
 		expectedTTL := 2 * time.Hour
 
@@ -108,15 +114,19 @@ func TestProviderManager_GetMeta(t *testing.T) {
 	})
 
 	t.Run("does not cache result with zero TTL and tries next provider", func(t *testing.T) {
-		zeroTTLMeta := pluginsv0alpha1.PluginMetaJSONData{
-			Id:   "test-plugin",
-			Name: "Zero TTL Plugin",
-			Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+		zeroTTLMeta := pluginsv0alpha1.PluginMetaSpec{
+			PluginJSON: pluginsv0alpha1.PluginMetaJSONData{
+				Id:   "test-plugin",
+				Name: "Zero TTL Plugin",
+				Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+			},
 		}
-		expectedMeta := pluginsv0alpha1.PluginMetaJSONData{
-			Id:   "test-plugin",
-			Name: "Test Plugin",
-			Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+		expectedMeta := pluginsv0alpha1.PluginMetaSpec{
+			PluginJSON: pluginsv0alpha1.PluginMetaJSONData{
+				Id:   "test-plugin",
+				Name: "Test Plugin",
+				Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+			},
 		}
 
 		provider1 := &mockProvider{
@@ -154,10 +164,12 @@ func TestProviderManager_GetMeta(t *testing.T) {
 	})
 
 	t.Run("tries next provider when first returns ErrMetaNotFound", func(t *testing.T) {
-		expectedMeta := pluginsv0alpha1.PluginMetaJSONData{
-			Id:   "test-plugin",
-			Name: "Test Plugin",
-			Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+		expectedMeta := pluginsv0alpha1.PluginMetaSpec{
+			PluginJSON: pluginsv0alpha1.PluginMetaJSONData{
+				Id:   "test-plugin",
+				Name: "Test Plugin",
+				Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+			},
 		}
 
 		provider1 := &mockProvider{
@@ -229,15 +241,19 @@ func TestProviderManager_GetMeta(t *testing.T) {
 	})
 
 	t.Run("skips expired cache entries", func(t *testing.T) {
-		expiredMeta := pluginsv0alpha1.PluginMetaJSONData{
-			Id:   "test-plugin",
-			Name: "Expired Plugin",
-			Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+		expiredMeta := pluginsv0alpha1.PluginMetaSpec{
+			PluginJSON: pluginsv0alpha1.PluginMetaJSONData{
+				Id:   "test-plugin",
+				Name: "Expired Plugin",
+				Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+			},
 		}
-		expectedMeta := pluginsv0alpha1.PluginMetaJSONData{
-			Id:   "test-plugin",
-			Name: "Test Plugin",
-			Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+		expectedMeta := pluginsv0alpha1.PluginMetaSpec{
+			PluginJSON: pluginsv0alpha1.PluginMetaJSONData{
+				Id:   "test-plugin",
+				Name: "Test Plugin",
+				Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+			},
 		}
 
 		callCount := 0
@@ -272,15 +288,19 @@ func TestProviderManager_GetMeta(t *testing.T) {
 	})
 
 	t.Run("uses first successful provider", func(t *testing.T) {
-		expectedMeta1 := pluginsv0alpha1.PluginMetaJSONData{
-			Id:   "test-plugin",
-			Name: "Provider 1 Plugin",
-			Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+		expectedMeta1 := pluginsv0alpha1.PluginMetaSpec{
+			PluginJSON: pluginsv0alpha1.PluginMetaJSONData{
+				Id:   "test-plugin",
+				Name: "Provider 1 Plugin",
+				Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+			},
 		}
-		expectedMeta2 := pluginsv0alpha1.PluginMetaJSONData{
-			Id:   "test-plugin",
-			Name: "Provider 2 Plugin",
-			Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+		expectedMeta2 := pluginsv0alpha1.PluginMetaSpec{
+			PluginJSON: pluginsv0alpha1.PluginMetaJSONData{
+				Id:   "test-plugin",
+				Name: "Provider 2 Plugin",
+				Type: pluginsv0alpha1.PluginMetaJSONDataTypeDatasource,
+			},
 		}
 
 		provider1 := &mockProvider{
@@ -331,9 +351,9 @@ func TestProviderManager_Run(t *testing.T) {
 
 func TestProviderManager_cleanupExpired(t *testing.T) {
 	t.Run("removes expired entries", func(t *testing.T) {
-		validMeta := pluginsv0alpha1.PluginMetaJSONData{Id: "valid"}
-		expiredMeta1 := pluginsv0alpha1.PluginMetaJSONData{Id: "expired1"}
-		expiredMeta2 := pluginsv0alpha1.PluginMetaJSONData{Id: "expired2"}
+		validMeta := pluginsv0alpha1.PluginMetaSpec{PluginJSON: pluginsv0alpha1.PluginMetaJSONData{Id: "valid"}}
+		expiredMeta1 := pluginsv0alpha1.PluginMetaSpec{PluginJSON: pluginsv0alpha1.PluginMetaJSONData{Id: "expired1"}}
+		expiredMeta2 := pluginsv0alpha1.PluginMetaSpec{PluginJSON: pluginsv0alpha1.PluginMetaJSONData{Id: "expired2"}}
 
 		provider := &mockProvider{
 			getMetaFunc: func(ctx context.Context, pluginID, version string) (*Result, error) {
