+++ pkg/server/wire_gen.go
@@ -788,7 +788,7 @@ func Initialize(ctx context.Context, cfg *setting.Cfg, opts Options, apiOpts api
 	if err != nil {
 		return nil, err
 	}
-	appInstaller, err := plugins.RegisterAppInstaller(configProvider, eventualRestConfigProvider)
+	appInstaller, err := plugins.RegisterAppInstaller(configProvider, eventualRestConfigProvider, pluginstoreService, pluginassetsService)
 	if err != nil {
 		return nil, err
 	}
@@ -1438,7 +1438,7 @@ func InitializeForTest(ctx context.Context, t sqlutil.ITestDB, testingT interfac
 	if err != nil {
 		return nil, err
 	}
-	appInstaller, err := plugins.RegisterAppInstaller(configProvider, eventualRestConfigProvider)
+	appInstaller, err := plugins.RegisterAppInstaller(configProvider, eventualRestConfigProvider, pluginstoreService, pluginassetsService)
 	if err != nil {
 		return nil, err
 	}
