+++ pkg/services/pluginsintegration/pluginstore/plugins.go
@@ -17,6 +17,7 @@ type Plugin struct {
 
 	// App fields
 	Parent          *ParentPlugin
+	Children        []string
 	IncludedInAppID string
 	DefaultNavURL   string
 	Pinned          bool
@@ -83,6 +84,18 @@ func ToGrafanaDTO(p *plugins.Plugin) Plugin {
 
 	if p.Parent != nil {
 		dto.Parent = &ParentPlugin{ID: p.Parent.ID}
+	}
+
+	if len(p.Children) > 0 {
+		children := make([]string, 0, len(p.Children))
+		for _, child := range p.Children {
+			if child != nil {
+				children = append(children, child.ID)
+			}
+		}
+		if len(children) > 0 {
+			dto.Children = children
+		}
 	}
 
 	return dto
